Add files via upload
DO NOT USE.

This current incarnation does NOT pull the pricing data correctly because I am censoring the results because of weird entries and blanks. I will implement a better mapping and zipping function like you find in my previous leaf.rb.

However, you can run this and be able to get results and know what is listed where, the pricing is just screwed up.

In addition, it's worth noting that this is a web scraping utilizing automation via ChromeDriver/Selenium/Chrome, so it will automate and launch multiple windows when going through.

There might be some errors that crop up because it can't extract text.

Please forgive how rudimentary it is.

Somehe5.csv is an example of the data created.